+++ SearchRobot.Library/Maps/Goal.cs
@@ -63,7 +63,7 @@ namespace SearchRobot.Library.Maps
 			canvas.Children.Remove(_uiElement);
 		}
 
-		public override void Move(Canvas canvas, int offsetX, int offsetY)
+		public override void Move(Canvas canvas, double offsetX, double offsetY)
 		{
 			StartPosition.X += offsetX;
 			StartPosition.Y += offsetY;
